+++ src/client/index.tsx
@@ -1,7 +1,9 @@
 import { render } from "preact";
 import { Route, Switch } from "wouter-preact";
 import "./styles.css";
+import { BucketProvider } from "./bucket";
 import { Layout } from "./components/Layout";
+import { useDocumentTitle } from "./hooks";
 import { DocByTitle } from "./pages/DocByTitle";
 import { DocDetail } from "./pages/DocDetail";
 import { DocEdit } from "./pages/DocEdit";
@@ -15,6 +17,7 @@ import { initTheme } from "./theme";
 initTheme();
 
 function NotFound() {
+  useDocumentTitle("404");
   return (
     <div class="page">
       <h1>404</h1>
@@ -25,19 +28,21 @@ function NotFound() {
 
 function App() {
   return (
-    <Layout>
-      <Switch>
-        <Route path="/" component={Home} />
-        <Route path="/docs" component={DocList} />
-        <Route path="/docs/title/:title">{(params) => <DocByTitle title={params.title} />}</Route>
-        <Route path="/docs/:key/edit">{(params) => <DocEdit docKey={params.key} />}</Route>
-        <Route path="/docs/:key">{(params) => <DocDetail docKey={params.key} />}</Route>
-        <Route path="/search" component={SearchPage} />
-        <Route path="/tags" component={TagsPage} />
-        <Route path="/graph" component={GraphPage} />
-        <Route component={NotFound} />
-      </Switch>
-    </Layout>
+    <BucketProvider>
+      <Layout>
+        <Switch>
+          <Route path="/" component={Home} />
+          <Route path="/docs" component={DocList} />
+          <Route path="/docs/title/:title">{(params) => <DocByTitle title={params.title} />}</Route>
+          <Route path="/docs/:key/edit">{(params) => <DocEdit docKey={params.key} />}</Route>
+          <Route path="/docs/:key">{(params) => <DocDetail docKey={params.key} />}</Route>
+          <Route path="/search" component={SearchPage} />
+          <Route path="/tags" component={TagsPage} />
+          <Route path="/graph" component={GraphPage} />
+          <Route component={NotFound} />
+        </Switch>
+      </Layout>
+    </BucketProvider>
   );
 }